Update test case with new markdown.py
